# Reduce execution time for skipping duplicate test_cases